Boundary snake (#115)
* Add Direction utility class

* Fix Vector.normalize

* Add Direction.getFromTo method

* Remove unnecessary check from WorldChunkCollection.findIntersectingChunks

* Basic boundary snake

* Tweaks

* Fix test

* Minor tweaks

+++ server/src/main/java/game/snake/GrowingSnakeChunk.java
@@ -32,7 +32,7 @@ public class GrowingSnakeChunk extends SnakeChunk {
 
         end = snake.headPosition.clone();
         pathData.add(new SnakePathPoint(this, end.clone(), 0.0));
-        endDirection = snake.headDirection;
+        endDirection = (float) snake.headDirection;
         direction = endDirection;
 
         minX = maxX = x = end.x;
